# swagger.yml
paths:
  /taxis:
    get:
      tags:
      - Taxi
      summary: Lista todos los taxis
      description: Listado de todos los taxis con su ID y placa
      operationId: getTaxis
      parameters:
      - in: query
        name: page
        description: Numero de pagina para la paginacion
        schema:
          type: integer
          format: int32
          minimum: 1
          default: 1
      - in: query
        name: per_page
        description: Taxis por pagina
        schema:
          type: integer
          format: int32
          minimum: 1
          default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxiSchema'
        '500':
          description: Error al obtener los datos de la base de datos
  /taxis/{taxi_id}/locations:
    get:
      tags:
      - Localizaciones
      summary: Lista de todas las localizaciones segun ID y fecha
      description: Listado de todas las localizaciones segun ID y fecha
      operationId: getTaxiLocations
      parameters:
      - in: path
        name: taxi_id
        description: ID del taxi
        required: true
        schema:
          type: integer
      - in: query
        name: date
        description: Fecha en formato YYYY-MM-DD
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Localizaciones del taxi obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaxiLocationSchema'
        '500':
          description: Error al obtener los datos de la base de datos
  /taxis/{taxi_id}/last-location:
    get:
      tags:
      - Ultima localizacion
      summary: Muestra la ultima localizacion de un taxi segun ID
      description: Listado de la ultima localizacion de un taxi segun ID
      operationId: getLastLocation
      parameters:
      - in: path
        name: taxi_id
        description: ID del taxi
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Ultima localizacion del taxi obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastLocationSchema'
        '404':
          description: No se encontraron ubicaciones para el taxi con ID proporcionado
        '500':
          description: Error al obtener los datos de la base de datos
info:
  title: Swagger Fleet Management API
  version: 1.0.0
openapi: 3.0.2
components:
  schemas:
    TaxiSchema:
      type: object
      properties:
        id:
          type: integer
          description: ID del taxi
        plate:
          type: string
          description: Placa del taxi
      required:
      - id
      - plate
    TaxiLocationSchema:
      type: object
      properties:
        fecha:
          type: string
          format: date
          description: Fecha de la ubicacion
        latitud:
          type: number
          description: Latitud de la ubicacion
        longitud:
          type: number
          description: Longitud de la ubicacion
      required:
      - fecha
      - latitud
      - longitud
    LastLocationSchema:
      type: object
      properties:
        id:
          type: integer
          description: ID del taxi
        fecha:
          type: string
          format: date
          description: Fecha de ultima localizacion
        latitud:
          type: number
          description: Latitud de la ultima localizacion
        longitud:
          type: number
          description: Longitud de la ultima localizacion
      required:
      - fecha
      - id
      - latitud
      - longitud